created a component for the deviceInfo
